refactor: move _networkx_to_dict
